Fix aws addition slowing down chown
Fixes slow chown-ing the working directory to the primary user caused by the AWS CLI being unpacked into the pwd.

The CLI is now unzipped into /tmp, installed, and then the installation files are removed.

While this was tested to ensure the AWS cli exists on the path, full functionality was not tested.
Removing the installation files can be omitted if the cli is found to still require them.